refactor(scripts): fixing unused imports
