+++ scripts/createLocal.js
@@ -1,7 +1,6 @@
 const { createAndExport } = require('@axelar-network/axelar-local-dev');
 const {
     Wallet,
-    utils: { keccak256, defaultAbiCoder },
 } = require('ethers');
 require('dotenv').config();
